pyproject.toml
[tool.poetry]
name = "scheduler-workshops"
version = "0.1.0"
description = ""
authors = ["KuligKamil <kamil.marek.kulig@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
pydantic = "^1.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
